+++ tests/test-cran/test-string.R
@@ -46,6 +46,11 @@ assert('split_source() puts lines of the same expression into a list element', {
   x = c('# a', '# b', '1', '# c', 'if (T)', 'F')
   (split_source(x) %==% c(as.list(x[1:4]), list(x[5:6])))
   (split_source(x, merge_comments = TRUE) %==% list(x[1:3], x[4:6]))
+  x = c(x, '1+1')
+  (split_source(x, TRUE, line_number = TRUE)[2:3] %==% list(
+    structure(x[4:6], line_start = 4),
+    structure(x[7], line_start = 7)
+  ))
 })
 
 assert('split_source() should signal an error for incomplete code', {